==> eDNA/02-datatable_prep_12S_Riaz.Rmd <==
---
title: "Datatable preparation base script for eDNA metabarcoding"
output:
  github_document: default
  pdf_document:
    keep_tex: yes
  html_document:
    toc: yes
    toc_depth: 6
    toc_float: yes
editor_options: 
  chunk_output_type: inline
---

This script takes your Blast output from the GMGI database, Mitofish database, and NCBI database to create one datatable with read counts and taxonomic assignment.  

**Workflow summary:**  
1. Load libraries  
2. Load metadata   
3. Load BLAST output from GMGI, Mitofish, and NCBI   
4. Load DADA2 ASV Table   
5. Taxonomic Assignment
- 5a. Identify ASVs with multiple hits from GMGI's database  
- 5b. Identify entries that mismatch between GMGI, Mitofish, and NCBI databases  
- 5c. Assign taxonomy based on hierarchical approach  
- 5d. Edit taxonomy annotations based on mismatch table choices  
- 5e. Adjusting common name and category for those entries that don't have one (from Mito or NCBI)  
6. Filtering: Filter ASV by less than 0.1% reads and then collapse by group  
7. Collapsing read counts by species name  
8. Creating results output 


# Load libraries

```{r}
library(ggplot2) ## for plotting
library(dplyr) ## for data table manipulation
library(tidyr) ## for data table manipulation
library(readr) ## for reading in tsv files
library(readxl) ## for reading in excel files
library(stringr) ## for data transformation
library(strex) ## for data transformation
library(writexl) ## for excel output
library(purrr) ## for data transformation
library(funrar) ## for make_relative()
library(tidyverse) ## for data table manipulation
```

# Metadata input

## Identify paths for metadata and project data 

Each user needs to write in their specific directory outputs prior to the file name. The default working directory is this document so the folder where this script is saved for the user. To change the workign directory to the Rproject directory, select 'Knit' and 'Knit Directory' > 'Project Directory'.

```{r}
### User edits:
### 1. change paths of input and output as desired 

## GMGI Fish database
path_GMGIdb = "../../../Projects/eDNA/Metabarcoding Lab Resources/Reference Databases/GMGI_Vert_Ref.xlsx"
path_fishbase_tax = "../../../Projects/eDNA/Metabarcoding Lab Resources/Reference Databases/taxonomic_classification_fishbase.csv"
path_mitofish_tax = "../../../Projects/eDNA/Metabarcoding Lab Resources/Reference Databases/taxonomic_classification_mitofish.csv"

## BLAST results
path_blast_gmgi = "example_input/BLASTResults_GMGI.txt"
path_blast_mito = "example_input/BLASTResults_Mito.txt"
path_blast_ncbi_taxassigned = "example_input/NCBI_taxassigned.txt"
path_blast_ncbi = "example_input/BLASTResults_NCBI.txt"

## ASV table results 
## confirm that the ASV_table.len.tsv name is correct for user's project
path_asv_table = "example_input/ASV_table.len.tsv"
path_output_summary = "example_input/overall_summary.tsv"

# output paths 
path_choice_required = "example_output/Taxonomic_assignments/Choice_required_GMGI_multiplehits.xlsx"
path_disagree_list = "example_output/Taxonomic_assignments/SampleReport_taxonomic_ID.xlsx"

results_rawreads_matrix = "example_output/Results_rawreads_matrix.xlsx"
results_rawreads_long = "example_output/Results_rawreads_long.xlsx"
results_relab_matrix = "example_output/Results_relab_matrix.xlsx"
results_relab_long = "example_output/Results_relab_long.xlsx"
```

## Load project metadata 

Metadata specific to each project. This contains information about each sample (e.g., month, site, time, sample type, etc.). Confirm that sample IDs match those used in the ASV_table.len.tsv file. 

```{r}
### User edits:
### 1. change path of metadata file

## EXCEL
# meta <- read_excel("example_input/metadata.xlsx")
## CSV 
meta <- read.csv("example_input/metadata.csv", header = TRUE)
```

## Load database metadata

No user edits in this section because paths have already been set above.

```{r}
# Load GMGI database information (common name, species name, etc.)
gmgi_db <- read_xlsx(path_GMGIdb, sheet = 1) %>% dplyr::rename(sseqid = Ref) %>%
  ## removing > from beginning of entires within Ref column
  mutate(sseqid = gsub(">", "", sseqid))
```

# BLAST data input 

No user edits unless user changed blastn parameters from fisheries team default.

```{r}
## Setting column header names and classes
blast_col_headers = c("ASV_ID", "sseqid", "pident", "length", "mismatch", "gapopen",
                                        "qstart", "qend", "sstart", "send", "evalue", "bitscore")
blast_col_classes = c(rep("character", 2), rep("numeric", 10))
```

## GMGI database

No user edits.

```{r}
Blast_GMGI <- read.table(path_blast_gmgi, header=F, col.names = blast_col_headers, colClasses = blast_col_classes) %>%
  ## blast changes spaces to hyphons so we need to change that back to match our metadata
  mutate(sseqid = gsub("-", " ", sseqid)) %>%
  ## join with GMGI database information
  left_join(., gmgi_db, by = "sseqid")

## Check how many ASVs were identified with the GMGI Database
length(unique(Blast_GMGI$ASV_ID)) 
```

## Mitofish database 

No user edits.

```{r}
Blast_Mito <- read.table(path_blast_mito, header=F, col.names = blast_col_headers, colClasses = blast_col_classes) %>%
  # renaming sseqid to species name
  dplyr::rename(Species_name = sseqid) %>%
  
  # replacing _ with spaces
  mutate(Species_name = gsub("_", " ", Species_name))
```


## NCBI database 

No user edits.

```{r}
NCBI_taxassigned <- read.delim2(path_blast_ncbi_taxassigned, header=F, col.names = c("staxid", "Phylo")) %>%
  ## creating taxonomic assignment columns
  separate(Phylo, c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species_name"), sep = ";") %>%
  ## creating species column based on Species_name
  mutate(., species = str_after_nth(Species_name, " ", 1))

Blast_NCBI <- read.table(path_blast_ncbi, header=F,
                           col.names = c("ASV_ID", "sseqid", "sscinames", "staxid", "pident", "length", "mismatch",
                                         "gapopen", "qstart", "qend", "sstart", "send", "evalue", "bitscore"),
                           colClasses = c(rep("character", 3), "integer", rep("numeric", 9))) %>%
  left_join(., NCBI_taxassigned, by = "staxid")
```


# Load DADA2 ASV Table 

The column headers will be the Sample IDs and the first column is the ASV ID. ASVs are given a "rank" based on sum of reads from that ASV (pre-filtering). 'Random' indicates that if ASVs are tied, then the code will randomly assign a rank for those tied. Because we don't need an exact rank here, 'random' will do for a tie-breaker.

No user edits.

```{r}
ASV_table <- read_tsv(path_asv_table, show_col_types = FALSE) %>%
  ## calculate the sum of all reads for each ASV
  mutate(., ASV_sum = rowSums(across(where(is.numeric)))) %>% 
  
  ## calculate a ranking based on those sum calculated above
  mutate(ASV_rank = rank(-ASV_sum, ties.method='random')) %>%
  
  ## move the sum and rank columns to after ASV_ID and arrange by rank
  relocate(c(ASV_sum,ASV_rank), .after = ASV_ID) %>% arrange((ASV_rank))

## creating list of rankings
ASV_rank_list <- ASV_table %>% dplyr::select(ASV_ID, ASV_sum, ASV_rank)
```

# Taxonomic Assignment

Identifying where NCBI, Mito, and GMGI disagree on tax assignment. With the hierarchial approach, ASVs that match to GMGI and several other databases will only result in GMGI assignment. By reviewing this df, we can be sure we aren't missing an assignment in our GMGI curated database. 

**Sub-workflow:**  
1. Identify any ASVs that contain multiple hits within the GMGI database.
2. Identify entries that mismatch between GMGI, Mitofish, and NCBI databases.
3. Assign taxonomy based on hierarchical approach.
4. Edit taxonomy annotations based on mismatch table.  
5. Adjusting common name for those entries that don't have one (from Mito or GMGI). 

## Identify any ASVs that contain multiple hits within the GMGI database

At this point, a fisheries team member needs to make choices about which taxonomic assignment to accept.

#### Create list of those ASVs with multiple hits

No user edits.

```{r}
multiple_hit_choice <- Blast_GMGI %>% group_by(ASV_ID) %>%
  ## take top percent identity hit, count the number of top hits, and filter to those with more than 1 top hit 
  slice_max(pident, n=1) %>% count() %>% filter(n>1) %>%
  
  ## adding BLAST_GMGI information with these ASVs and ASV rank and sum
  left_join(., Blast_GMGI, by = "ASV_ID") %>%
  left_join(., ASV_rank_list, by = "ASV_ID") %>%
  
  ## moving database percent ID to be next to Blast percent ID
  relocate(c(db_percent_ID, ASV_sum, ASV_rank), .after = pident) %>%
  
  ## adding choice column for next steps 
  mutate(Choice = NA)

## export this data frame as excel sheet 
multiple_hit_choice %>% write_xlsx(path_choice_required)
```

Based on the output above, user needs to make some choices. In the excel spreadsheet, user needs to mark 'x' on the choices desired while leaving the other entries blank. 

#### Choosing one of several hits. 

Load choice edited dataset. No user edits. 

```{r}
multiple_hit_choice_edited <- read_xlsx("example_output/Taxonomic_assignments/Choice_required_GMGI_multiplehits_edited.xlsx") %>%
  ## selecting the choices made
  filter(!is.na(Choice)) %>%
  ## selecting only columns needed 
  dplyr::select(ASV_ID, sseqid, Choice)
```

A for loop will filter Blast_GMGI df based on these choices. No user edits. 

```{r}
# Create a new edited df
Blast_GMGI_edited <- Blast_GMGI 

# Loop through each row of the dataframe
for (i in multiple_hit_choice_edited$ASV_ID) {
  # Extract the current row (will do this for each ASV_ID in the choice df)
  current_row <- multiple_hit_choice_edited %>% subset(ASV_ID==i)
  
  # Apply filter based on the current row's condition
  Blast_GMGI_edited <- Blast_GMGI_edited %>%
    filter(case_when(ASV_ID == current_row$ASV_ID ~ sseqid == current_row$sseqid,
           TRUE ~ TRUE))
}
```

#### Confirming that all entries have been dealth with

No user edits.

```{r}
### Check the below output to confirm the filtering steps above worked (if it worked, it won't be in output)
Blast_GMGI_edited %>% group_by(ASV_ID) %>% slice_max(pident, n=1) %>% count() %>% filter(n>1)
```


## Identify entries that mismatch between GMGI, Mitofish, and NCBI databases

Creating a df called "Disagree". Review the output before moving onto the next section.

No user edits.

```{r}
Disagree <- Blast_GMGI_edited %>% group_by(ASV_ID) %>% 
  dplyr::rename(., GMGI_db_ID = db_percent_ID, GMGI_pident = pident) %>%
  ## Creating new columns with species name based on pident information
  mutate(
    GMGI_100 = if_else(GMGI_pident == 100, Species_name, NA_character_),
    GMGI_lessthan100 = if_else(GMGI_pident < 100, Species_name, NA_character_)) %>%
  
  ## taking only the top hit per ASV ID
  slice_max(GMGI_pident, n = 1, with_ties = FALSE) %>% ungroup() %>%

  ## filtering to distinct rows with selected columns
  distinct(ASV_ID, GMGI_db_ID, GMGI_pident, GMGI_100, GMGI_lessthan100) %>%
  
  ## adding Mitofish and editing the Blast_Mito df in the process
  full_join(Blast_Mito %>% dplyr::select(ASV_ID, Species_name) %>%
              dplyr::rename(Mitofish = Species_name) %>%
              distinct() %>% group_by(ASV_ID) %>%
              mutate(Mitofish = paste0(Mitofish, collapse = ";")),
            by = "ASV_ID") %>%
  
  ## adding NCBI and editing the Blast_NCBI df in the process
  full_join(Blast_NCBI %>% dplyr::select(ASV_ID, Species_name) %>%
              dplyr::rename(NCBI = Species_name) %>%
              distinct() %>% group_by(ASV_ID) %>%
              mutate(NCBI = paste0(NCBI, collapse = ";")),
            by = "ASV_ID") %>%
  
  ## adding ASV rank and sum information
  left_join(., ASV_rank_list, by = "ASV_ID") %>%

  ## filtering out duplicate rows
  distinct() %>%
  
  ## filtering to those entries that mismatch between GMGI, Mitofish, and NCBI
  filter((GMGI_100 != GMGI_lessthan100 | GMGI_100 != Mitofish | GMGI_100 != NCBI | is.na(GMGI_100))) %>%
  
  ## adding choice column for next steps 
  mutate(Choice = NA)

## export this data frame as excel sheet 
Disagree %>% write_xlsx(path_disagree_list)
```

## Assign taxonomy based on hierarchical approach

Taxonomic identification is taken from GMGI 100%, then GMGI <100%, then Mitofish 100%, and finally NCBI 100%.

No user edits.

```{r}
ASV_table_taxID <- ASV_table %>% 
  
  ## 1. Top hit from GMGI's database
  left_join(Blast_GMGI_edited %>%  group_by(ASV_ID) %>%
              slice_max(pident, n = 1) %>%
                            dplyr::select(ASV_ID, Species_name),
            by = join_by(ASV_ID)) %>%
  
  ## 2. Mitofish database
  ### join df, select ASV_ID and Species_name columns, rename Species_name to Mito, call only distinct rows
  left_join(., Blast_Mito %>% dplyr::select(ASV_ID, Species_name) %>% dplyr::rename(Mito = Species_name) %>% distinct() %>%
              
              ### group by ASV_ID, and collapse all species names separated by ;, then take only distinct rows
              group_by(ASV_ID) %>% mutate(Mito = paste0(Mito, collapse = ";")) %>% distinct(), by = "ASV_ID") %>%
  
  ### if GMGI annotation is NA, then replace with Mitofish 
  mutate(., Species_name = ifelse(is.na(Species_name), Mito, Species_name)) %>%

  ## 3. NCBI database; same functions as above
  left_join(., Blast_NCBI %>% dplyr::select(ASV_ID, Species_name) %>% dplyr::rename(NCBI = Species_name) %>% distinct() %>%
              group_by(ASV_ID) %>% mutate(NCBI = paste0(NCBI, collapse = ";")) %>% distinct(), by = "ASV_ID") %>%
  mutate(., Species_name = ifelse(is.na(Species_name), NCBI, Species_name)) %>%
  
  ## 4. if Species name is STILL not filled, call it "unassigned"
  mutate(., Species_name = ifelse(is.na(Species_name), "unassigned", Species_name)) %>%  

  ## removing Mito spp and NCBI spp
  dplyr::select(-Mito, -NCBI) %>%
  
  ## move species name to be after ASV_ID
  relocate(., c(Species_name), .after = ASV_ID)
```

## Edit taxonomy annotations based on mismatch table

Override any annotations with edited taxonomic identification table.   
No user edits.

```{r}
## read in edited df 
taxonomic_choice <- read_xlsx("data/BLASToutput/SampleReport_taxonomic_ID_edited.xlsx") %>%
    ## selecting only columns needed 
  dplyr::select(ASV_ID, Choice)  

# Create a new edited df
ASV_table_taxID_edited <- ASV_table_taxID 

# Loop through each row of the dataframe
for (i in taxonomic_choice$ASV_ID) {
  # Extract the current row (will do this for each ASV_ID in the choice df)
  current_row <- taxonomic_choice %>% subset(ASV_ID==i)
  
  # Apply filter based on the current row's condition
  ASV_table_taxID_edited <- ASV_table_taxID_edited %>%
    mutate(Species_name = case_when(
          ASV_ID == current_row$ASV_ID ~ current_row$Choice,
           TRUE ~ Species_name))
}
```

### Confirm all entries are dealt with

No user edits.

```{r}
## Output will be blank
ASV_table_taxID_edited %>% dplyr::select(Species_name) %>% distinct() %>% 
  filter(., grepl(";", Species_name)) %>% arrange(Species_name) 
```

## Adjusting common name for those entries that don't have one (from Mito or NCBI)

No user edits.

```{r}
### add common name column to df
ASV_table_taxID_edited <- ASV_table_taxID_edited %>%
  left_join(., gmgi_db %>% dplyr::select(Species_name, Common_name, Category) %>% distinct(), by = "Species_name") %>%
  relocate(., c(Common_name, Category), .after = Species_name)

### print entries with no common name
ASV_table_taxID_edited %>% dplyr::select(Species_name, Common_name) %>% filter(is.na(Common_name)) %>% distinct()
```

Editing common names and category when needed.

```{r}
### User edits:
### 1. Add mutate cases using the format ifelse(grepl('', Species_name), "", Common_name
### example: ifelse(grepl('unassigned', Species_name), "unassigned", Common_name)
### 2. Add mutate cases using the format ifelse(grepl('', Species_name), "", Category

ASV_table_taxID_edited <- ASV_table_taxID_edited %>% 
  # changing specific entries for Common name
  mutate(Common_name = ifelse(grepl('unassigned', Species_name), "unassigned", Common_name),
         
         ## example from example dataset - change for your own data
         Common_name = ifelse(grepl('Cololabis saira', Species_name), "Pacific Saury", Common_name)
         ) %>%
  
  # changing specific entries for category
  mutate(Category = ifelse(grepl('unassigned', Species_name), "unassigned", Category),
         
         ## example from example dataset - change for your own data
         Category = ifelse(grepl('Cololabis saira', Species_name), "Teleost Fish", Category)
         )

## printing list of species name without common names 
## after additions to mutate function above, this output should be zero 
ASV_table_taxID_edited %>% dplyr::select(Species_name, Common_name) %>% filter(is.na(Common_name)) %>% distinct()
```

# Filtering: Filter ASV by less than 0.1% reads and then collapse by group

## Filter out reads that are less than 0.1% of ASV (row) total per sample. 

Create an output of what you're losing with filtering. 

No user edits.

```{r}
ASV_table_taxID_filtered <- ASV_table_taxID_edited %>%
  ## telling the df we are doing the following function by rows (ASVs)
  rowwise() %>%
  
  ## filtering out any values that are less than 0.001 of the total ASV read # in each sample
  mutate(across(.cols = (7:ncol(.)),            
                .fns = ~ ifelse((.x/ASV_sum)<0.001, NA, .x))) %>% ungroup()

## output of what we're losing
ASV_table_taxID_edited %>% rowwise() %>%
  mutate(across(.cols = (7:ncol(.)),            
                .fns = ~ ifelse((.x/ASV_sum)>0.001, NA, .x))) %>% ungroup() %>% write_xlsx("data/ASV_reads_filtered_out.xlsx")


## Export ASV break-down for 03-data_quality.Rmd
ASV_table_taxID_filtered %>% dplyr::select(ASV_ID, Species_name, Common_name, Category, ASV_sum, ASV_rank) %>%
  write_xlsx("data/ASV_breakdown.xlsx")
```


# Collapsing read counts by species name

No user edits.

```{r}
ASV_table_taxID_collapsed <- ASV_table_taxID_filtered %>% 
  # removing original ASV_ID to collapse
  dplyr::select(-ASV_ID) %>%  
  
  ## group by Species_name and sample
  dplyr::group_by(Species_name, Common_name, Category) %>%
  
  ## sum down column by species name and sample to collapse
  summarise(across(6:last_col(), ~ sum(., na.rm = TRUE))) %>% ungroup()
```


# Creating results output 

Raw reads results output.   
No user edits.  

```{r}
## Raw reads matrix (wide format)
ASV_table_taxID_collapsed %>% write_xlsx(results_rawreads_matrix)

## Raw reads long format and filtering out entries with zero reads
ASV_table_taxID_collapsed %>% 
  gather("sampleID", "reads", c(4:last_col())) %>%
  filter(reads > 0) %>% 
  left_join(., meta, by = "sampleID") %>%
  write_xlsx(results_rawreads_long)
```

Relative Abundance

```{r}
## Creating matrix from edited collapsed df
data.matrix <- as.matrix(ASV_table_taxID_collapsed %>% 
                           dplyr::select(-Common_name, -Category) %>% column_to_rownames(var = "Species_name"))

## Calculating relative abundance
data_relativeab <- as.data.frame(make_relative(data.matrix)) %>%
  ## moving rownames to a column
  rownames_to_column(var = "Species_name") %>%
  
  ## adding common name and category back in
  right_join(ASV_table_taxID_collapsed %>% dplyr::select(Species_name, Common_name, Category), .)  
  
## Exporting matrix
data_relativeab %>% write_xlsx(results_relab_matrix)

## Relative abundance long format with metadata 
data_relativeab %>%
  gather("sampleID", "relab", c(4:last_col())) %>%
  left_join(., meta, by = "sampleID") %>%
  write_xlsx(results_relab_long)
```
